Resolved some warnings

+++ com.openexchange.admin.reseller/src/com/openexchange/admin/reseller/rmi/impl/ResellerAuth.java
@@ -104,7 +104,7 @@ public class ResellerAuth extends OXCommonImpl implements BasicAuthenticatorPlug
                 return false;
             }
             ResellerAdmin adm = oxresell.getData(new ResellerAdmin[]{new ResellerAdmin(creds.getLogin())})[0];
-            return oxresell.ownsContextOrIsPidOfOwner(ctx, adm.getId());
+            return oxresell.ownsContextOrIsPidOfOwner(ctx, adm.getId().intValue());
         } catch (StorageException e) {
             log.error("",e);
             throw new InvalidCredentialsException("authentication failed");
